edit recipe + update schema
