+++ src/app/admin/page.tsx
@@ -1,11 +1,11 @@
 import { Button } from '@/components/ui/button';
 import Link from 'next/link';
-import RecipesList from '@/components/admin/RecipesList';
+import RecipesList from '@/components/admin/recipes/RecipesList';
 
 const AdminPage = () => {
   return (
     <div className="p-14 w-full">
-      <Link href="recipes/new">
+      <Link href="admin/recipes/new">
         <Button size="lg">Create new recipe</Button>
       </Link>
       <RecipesList />
