+++ src/app/api/ingredients/route.ts
@@ -1,6 +1,5 @@
 import { NextRequest } from 'next/server';
 import prisma from '@/lib/prisma';
-import { revalidatePath } from 'next/cache';
 
 export async function POST(request: NextRequest) {
   const json: {
@@ -33,8 +32,6 @@ export async function POST(request: NextRequest) {
       name: name.charAt(0).toUpperCase() + name.slice(1),
     },
   });
-
-  revalidatePath('/admin');
 
   return Response.json(created, {
     status: 201,
