+++ src/app/api/recipes/[id]/route.ts
@@ -1,12 +1,15 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { createSchema } from '@/utils/recipes/form.utils';
 import { auth } from '@/auth';
+import { createSchema } from '@/utils/recipes/form.utils';
 import prisma from '@/lib/prisma';
 import { revalidatePath } from 'next/cache';
 
-export async function POST(request: NextRequest) {
+export async function PUT(
+  request: NextRequest,
+  { params }: { params: Promise<{ id: string }> }
+) {
   const session = await auth();
-
+  const id = parseInt((await params).id);
   if (!session?.user) {
     return NextResponse.json({ error: 'User not found' }, { status: 403 });
   }
@@ -16,19 +19,17 @@ export async function POST(request: NextRequest) {
     const { title, type, ingredients, description, category } =
       createSchema.parse(json);
 
-    const response = await prisma.recipe.create({
+    const response = await prisma.recipe.update({
+      where: {
+        id,
+      },
       data: {
         title,
         type,
         description,
         category,
-        author: {
-          connect: {
-            email: session.user.email,
-          },
-        },
         ingredients: {
-          connect: ingredients.map((ingredient) => ({
+          set: ingredients.map((ingredient) => ({
             id: ingredient.id,
           })),
         },
@@ -36,9 +37,10 @@ export async function POST(request: NextRequest) {
     });
 
     revalidatePath('/admin');
+    revalidatePath(`/admin/recipes/${id}`);
 
     return NextResponse.json(response, {
-      status: 201,
+      status: 200,
     });
   } catch (e) {
     return NextResponse.json({ e }, { status: 400 });
